requirements.txt
cbor2==5.6.4
pyserial==3.5
pyelftools==0.31
setuptools==70.0.0
matplotlib
